+++ tdrs-backend/tdpservice/data_files/models.py
@@ -13,6 +13,8 @@ from django.db.models import Max
 from tdpservice.backends import DataFilesS3Storage
 from tdpservice.stts.models import STT
 from tdpservice.users.models import User
+from tdpservice.data_files.s3_client import S3Client
+from django.conf import settings
 
 logger = logging.getLogger(__name__)
 
@@ -150,6 +152,16 @@ class DataFile(FileRecord):
                                blank=False,
                                null=True
                                )
+    
+    @property
+    def download_file(self):
+        """download the file associated with version."""
+        s3_client=S3Client().client
+        # get object with version id that is associated with this field
+        bucket_name = settings.AWS_S3_DATAFILES_BUCKET_NAME
+        obj = s3_client.get_object(Bucket=bucket_name, Key=self.file.name, VersionId=self.s3_versioning_id)
+        print(obj)
+
 
     @property
     def filename(self):
